Move enchant modules into more appropiate location
