+++ src/parser/core/CombatLogParser.tsx
@@ -88,19 +88,19 @@ import { PetInfo } from './Pet';
 import { PlayerInfo } from './Player';
 import Report from './Report';
 import { SpellUsageContextProvider } from 'parser/core/SpellUsage/core';
-import BurningDevotion from 'parser/retail/modules/features/enchants/dragonflight/BurningDevotion';
-import BurningWrit from 'parser/retail/modules/features/enchants/dragonflight/BurningWrit';
-import EarthenDevotion from 'parser/retail/modules/features/enchants/dragonflight/EarthenDevotion';
-import EarthenWrit from 'parser/retail/modules/features/enchants/dragonflight/EarthenWrit';
-import FrozenDevotion from 'parser/retail/modules/features/enchants/dragonflight/FrozenDevotion';
-import FrozenWrit from 'parser/retail/modules/features/enchants/dragonflight/FrozenWrit';
-import InvigoratingSporeCloud from 'parser/retail/modules/features/enchants/dragonflight/InvigoratingSporeCloud';
-import ShadowflameWreathe from 'parser/retail/modules/features/enchants/dragonflight/ShadowflameWreathe';
-import SophicDevotion from 'parser/retail/modules/features/enchants/dragonflight/SophicDevotion';
-import SophicWrit from 'parser/retail/modules/features/enchants/dragonflight/SophicWrit';
-import SporeTender from 'parser/retail/modules/features/enchants/dragonflight/SporeTender';
-import WaftingDevotion from 'parser/retail/modules/features/enchants/dragonflight/WaftingDevotion';
-import WaftingWrit from 'parser/retail/modules/features/enchants/dragonflight/WaftingWrit';
+import BurningDevotion from 'parser/retail/modules/items/dragonflight/enchants/BurningDevotion';
+import BurningWrit from 'parser/retail/modules/items/dragonflight/enchants/BurningWrit';
+import EarthenDevotion from 'parser/retail/modules/items/dragonflight/enchants/EarthenDevotion';
+import EarthenWrit from 'parser/retail/modules/items/dragonflight/enchants/EarthenWrit';
+import FrozenDevotion from 'parser/retail/modules/items/dragonflight/enchants/FrozenDevotion';
+import FrozenWrit from 'parser/retail/modules/items/dragonflight/enchants/FrozenWrit';
+import InvigoratingSporeCloud from 'parser/retail/modules/items/dragonflight/enchants/InvigoratingSporeCloud';
+import ShadowflameWreathe from 'parser/retail/modules/items/dragonflight/enchants/ShadowflameWreathe';
+import SophicDevotion from 'parser/retail/modules/items/dragonflight/enchants/SophicDevotion';
+import SophicWrit from 'parser/retail/modules/items/dragonflight/enchants/SophicWrit';
+import SporeTender from 'parser/retail/modules/items/dragonflight/enchants/SporeTender';
+import WaftingDevotion from 'parser/retail/modules/items/dragonflight/enchants/WaftingDevotion';
+import WaftingWrit from 'parser/retail/modules/items/dragonflight/enchants/WaftingWrit';
 import AcceleratingSandglass from 'parser/retail/modules/items/dragonflight/AcceleratingSandglass';
 import VoiceOfTheSilentStar from 'parser/retail/modules/items/dragonflight/VoiceOfTheSilentStar';
