+++ src/parser/retail/modules/items/dragonflight/enchants/BurningDevotion.tsx
