+++ src/parser/retail/modules/items/dragonflight/enchants/BurningWrit.tsx
