+++ src/parser/retail/modules/items/dragonflight/enchants/EarthenWrit.tsx
