+++ src/parser/retail/modules/items/dragonflight/enchants/FrozenDevotion.tsx
