+++ src/parser/retail/modules/items/dragonflight/enchants/FrozenWrit.tsx
