+++ src/parser/retail/modules/items/dragonflight/enchants/InvigoratingSporeCloud.tsx
