+++ src/parser/retail/modules/items/dragonflight/enchants/SophicDevotion.tsx
