+++ src/parser/retail/modules/items/dragonflight/enchants/SophicWrit.tsx
