+++ src/parser/retail/modules/items/dragonflight/enchants/SporeTender.tsx
