+++ src/parser/retail/modules/items/dragonflight/enchants/WaftingWrit.tsx
